% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment7.R
\docType{package}
\name{alignment7}
\alias{alignment7}
\alias{alignment7-package}
\title{Alignment 7 experiment}
\description{
Data and materials for the Alignment 7 experiment.
}
\details{
alignment7
A data package for the Alignment 7 experiment.

Use \code{data(package='alignment7')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{data}
}
