% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment7.R
\docType{data}
\name{data}
\alias{data}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{id}{}
\item{subject}{}
\item{sex}{}
\item{age}{}
\item{other}{}
\item{business_exp}{}
\item{business_edu}{}
\item{project}{}
\item{alignment}{}
\item{reliability_type}{}
\item{reliability_amount}{}
\item{npv_cond}{}
\item{npv_amount}{}
\item{total_time}{}
\item{order}{}
\item{ranking}{}
\item{allocation}{}
\item{reliability_type_order}{}
\item{alignment_order}{}
\item{sample}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{alignment7}
}
